+++ QueenCam/QueenCam/Presentation/Views/Camera/UIComponent/CameraBottomContainer/CameraBottomContainer+ThumbnailButton.swift
@@ -8,13 +8,13 @@ extension CameraView.CameraBottomContainer {
       if let image = cameraViewModel.lastImage {
         Image(uiImage: image)
           .resizable()
-          .frame(width: 48, height: 48)
+          .frame(width: isMinimize ? 24 : 48, height: isMinimize ? 24 : 48)
           .clipShape(Circle())
       } else {
         if let thumbnailImage = cameraViewModel.thumbnailImage {
           Image(uiImage: thumbnailImage)
             .resizable()
-            .frame(width: 48, height: 48)
+            .frame(width: isMinimize ? 24 : 48, height: isMinimize ? 24 : 48)
             .clipShape(Circle())
         } else {
           EmptyPhotoButton()